+++ lib/searchkick/model.rb
@@ -2,7 +2,7 @@ module Searchkick
   module Model
     def searchkick(**options)
       unknown_keywords = options.keys - [:_all, :_type, :batch_size, :callbacks, :conversions, :default_fields,
-        :filterable, :geo_shape, :highlight, :ignore_above, :index_name, :index_prefix, :language,
+        :filterable, :geo_shape, :highlight, :ignore_above, :index_name, :index_prefix, :inheritance, :language,
         :locations, :mappings, :match, :merge_mappings, :routing, :searchable, :settings, :similarity,
         :special_characters, :stem_conversions, :suggest, :synonyms, :text_end,
         :text_middle, :text_start, :word, :wordnet, :word_end, :word_middle, :word_start]
@@ -11,6 +11,8 @@ module Searchkick
       raise "Only call searchkick once per model" if respond_to?(:searchkick_index)
 
       Searchkick.models << self
+
+      options[:_type] ||= -> { searchkick_index.klass_document_type(self) } if options[:inheritance]
 
       class_eval do
         cattr_reader :searchkick_options, :searchkick_klass